Fixed 'no logging' and 'noise.h'.

+++ depends/goom-libs/src/goom/src/utils/mathutils.cpp
@@ -154,7 +154,9 @@ SineWaveMultiplier::SineWaveMultiplier(const float freq,
 
 auto SineWaveMultiplier::GetNext() -> float
 {
-  const auto val = static_cast<float>(m_rangeMapper(m_lower, m_upper, std::sin(m_frequency * m_x)));
+  const auto val =
+      static_cast<float>(m_rangeMapper(static_cast<double>(m_lower), static_cast<double>(m_upper),
+                                       static_cast<double>(std::sin(m_frequency * m_x))));
   //logInfo("lower = {}, upper = {}, sin({}) = {}.", lower, upper, x, val);
   m_x += m_piStepFrac * m_pi;
   return val;
